+++ 2DAClike/Circle.h
@@ -8,7 +8,7 @@ public:
 	static bool Intersect(const Circle& r1, const Circle& r2);
 	void DrawDebug(const VECTOR& objectpos)
 	{
-		DrawCircle(position.x, position.y, (0, 255, 0), FALSE, 1);
+		DrawCircle(static_cast<int>(position.x), static_cast<int>(position.y), (0, 255, 0), FALSE, 1);
 	}
 };
